pgsql: Split feature.go to table based files in feature module
