+++ database/pgsql/feature/feature_test.go
@@ -12,36 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pgsql
+package feature
 
 import (
+	"database/sql"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/coreos/clair/database"
+	"github.com/coreos/clair/database/pgsql/testutil"
 )
 
 func TestPersistFeatures(t *testing.T) {
-	tx, cleanup := createTestPgSession(t, "TestPersistFeatures")
+	tx, cleanup := testutil.CreateTestTx(t, "TestPersistFeatures")
 	defer cleanup()
 
 	invalid := database.Feature{}
 	valid := *database.NewBinaryPackage("mount", "2.31.1-0.4ubuntu3.1", "dpkg")
 
 	// invalid
-	require.NotNil(t, tx.PersistFeatures([]database.Feature{invalid}))
+	require.NotNil(t, PersistFeatures(tx, []database.Feature{invalid}))
 	// existing
-	require.Nil(t, tx.PersistFeatures([]database.Feature{valid}))
-	require.Nil(t, tx.PersistFeatures([]database.Feature{valid}))
+	require.Nil(t, PersistFeatures(tx, []database.Feature{valid}))
+	require.Nil(t, PersistFeatures(tx, []database.Feature{valid}))
 
 	features := selectAllFeatures(t, tx)
 	assert.Equal(t, []database.Feature{valid}, features)
 }
 
 func TestPersistNamespacedFeatures(t *testing.T) {
-	tx, cleanup := createTestPgSessionWithFixtures(t, "TestPersistNamespacedFeatures")
+	tx, cleanup := testutil.CreateTestTxWithFixtures(t, "TestPersistNamespacedFeatures")
 	defer cleanup()
 
 	// existing features
@@ -58,42 +60,17 @@ func TestPersistNamespacedFeatures(t *testing.T) {
 	nf2 := database.NewNamespacedFeature(n2, f2)
 	// namespaced features with namespaces or features not in the database will
 	// generate error.
-	assert.Nil(t, tx.PersistNamespacedFeatures([]database.NamespacedFeature{}))
-	assert.NotNil(t, tx.PersistNamespacedFeatures([]database.NamespacedFeature{*nf1, *nf2}))
+	assert.Nil(t, PersistNamespacedFeatures(tx, []database.NamespacedFeature{}))
+	assert.NotNil(t, PersistNamespacedFeatures(tx, []database.NamespacedFeature{*nf1, *nf2}))
 	// valid case: insert nf3
-	assert.Nil(t, tx.PersistNamespacedFeatures([]database.NamespacedFeature{*nf1}))
+	assert.Nil(t, PersistNamespacedFeatures(tx, []database.NamespacedFeature{*nf1}))
 
 	all := listNamespacedFeatures(t, tx)
 	assert.Contains(t, all, *nf1)
 }
 
-func TestFindAffectedNamespacedFeatures(t *testing.T) {
-	datastore, tx := openSessionForTest(t, "FindAffectedNamespacedFeatures", true)
-	defer closeTest(t, datastore, tx)
-	ns := database.NamespacedFeature{
-		Feature: database.Feature{
-			Name:          "openssl",
-			Version:       "1.0",
-			VersionFormat: "dpkg",
-			Type:          database.SourcePackage,
-		},
-		Namespace: database.Namespace{
-			Name:          "debian:7",
-			VersionFormat: "dpkg",
-		},
-	}
-
-	ans, err := tx.FindAffectedNamespacedFeatures([]database.NamespacedFeature{ns})
-	if assert.Nil(t, err) &&
-		assert.Len(t, ans, 1) &&
-		assert.True(t, ans[0].Valid) &&
-		assert.Len(t, ans[0].AffectedBy, 1) {
-		assert.Equal(t, "CVE-OPENSSL-1-DEB7", ans[0].AffectedBy[0].Name)
-	}
-}
-
-func listNamespacedFeatures(t *testing.T, tx *pgSession) []database.NamespacedFeature {
-	types, err := tx.getFeatureTypeMap()
+func listNamespacedFeatures(t *testing.T, tx *sql.Tx) []database.NamespacedFeature {
+	types, err := GetFeatureTypeMap(tx)
 	if err != nil {
 		panic(err)
 	}
@@ -114,15 +91,15 @@ func listNamespacedFeatures(t *testing.T, tx *pgSession) []database.NamespacedFe
 			panic(err)
 		}
 
-		f.Type = types.byID[typeID]
+		f.Type = types.ByID[typeID]
 		nf = append(nf, f)
 	}
 
 	return nf
 }
 
-func selectAllFeatures(t *testing.T, tx *pgSession) []database.Feature {
-	types, err := tx.getFeatureTypeMap()
+func selectAllFeatures(t *testing.T, tx *sql.Tx) []database.Feature {
+	types, err := GetFeatureTypeMap(tx)
 	if err != nil {
 		panic(err)
 	}
@@ -137,7 +114,7 @@ func selectAllFeatures(t *testing.T, tx *pgSession) []database.Feature {
 		f := database.Feature{}
 		var typeID int
 		err := rows.Scan(&f.Name, &f.Version, &f.VersionFormat, &typeID)
-		f.Type = types.byID[typeID]
+		f.Type = types.ByID[typeID]
 		if err != nil {
 			t.FailNow()
 		}
@@ -146,45 +123,24 @@ func selectAllFeatures(t *testing.T, tx *pgSession) []database.Feature {
 	return fs
 }
 
-func assertNamespacedFeatureEqual(t *testing.T, expected []database.NamespacedFeature, actual []database.NamespacedFeature) bool {
-	if assert.Len(t, actual, len(expected)) {
-		has := map[database.NamespacedFeature]bool{}
-		for _, nf := range expected {
-			has[nf] = false
-		}
-
-		for _, nf := range actual {
-			has[nf] = true
-		}
-
-		for nf, visited := range has {
-			if !assert.True(t, visited, nf.Namespace.Name+":"+nf.Name+" is expected") {
-				return false
-			}
-		}
-		return true
-	}
-	return false
-}
-
 func TestFindNamespacedFeatureIDs(t *testing.T) {
-	tx, cleanup := createTestPgSessionWithFixtures(t, "TestFindNamespacedFeatureIDs")
+	tx, cleanup := testutil.CreateTestTxWithFixtures(t, "TestFindNamespacedFeatureIDs")
 	defer cleanup()
 
 	features := []database.NamespacedFeature{}
 	expectedIDs := []int{}
-	for id, feature := range realNamespacedFeatures {
+	for id, feature := range testutil.RealNamespacedFeatures {
 		features = append(features, feature)
 		expectedIDs = append(expectedIDs, id)
 	}
 
-	features = append(features, realNamespacedFeatures[1]) // test duplicated
+	features = append(features, testutil.RealNamespacedFeatures[1]) // test duplicated
 	expectedIDs = append(expectedIDs, 1)
 
-	namespace := realNamespaces[1]
+	namespace := testutil.RealNamespaces[1]
 	features = append(features, *database.NewNamespacedFeature(&namespace, database.NewBinaryPackage("not-found", "1.0", "dpkg"))) // test not found feature
 
-	ids, err := tx.findNamespacedFeatureIDs(features)
+	ids, err := FindNamespacedFeatureIDs(tx, features)
 	require.Nil(t, err)
 	require.Len(t, ids, len(expectedIDs)+1)
 	for i, id := range ids {
